requirements.txt
streamlit
python-dotenv
groq
opencv-python
numpy
mediapipe
Pillow
pyttsx3
pygame